[refactor] book adapter안의 repository 패키지 분리 (#80)
